+++ packages/core/src/bases/chain-wallet.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-expressions */
 import {
   CosmWasmClient,
   SigningCosmWasmClient,
@@ -31,14 +32,17 @@ import {
   getNameServiceRegistryFromChainName,
   isValidEndpoint,
 } from '../utils';
+import { MainWalletBase } from './main-wallet';
 import { WalletBase } from './wallet';
 
 export class ChainWalletBase extends WalletBase {
+  mainWallet: MainWalletBase;
   protected _chainRecord: ChainRecord;
   rpcEndpoints?: (string | ExtendedHttpEndpoint)[] = [];
   restEndpoints?: (string | ExtendedHttpEndpoint)[] = [];
   protected _rpcEndpoint?: string | ExtendedHttpEndpoint;
   protected _restEndpoint?: string | ExtendedHttpEndpoint;
+  connectChains?: () => Promise<any>;
   offlineSigner?: OfflineSigner;
   namespace = 'cosmos';
   isLazy?: boolean; // stands for real `chainIsLazy` considered both `globalIsLazy` and `chainIsLazy` settings
@@ -146,12 +150,16 @@ export class ChainWalletBase extends WalletBase {
     throw new Error('initClient not implemented');
   }
 
-  async update() {
+  async update(options = { connect: true }) {
     this.setState(State.Pending);
     this.setMessage(void 0);
 
     try {
-      await this?.client?.connect?.(this.chainId);
+      if (options.connect) {
+        this.connectChains
+          ? await this.connectChains()
+          : await this?.client?.connect?.(this.chainId);
+      }
 
       let account: SimpleAccount;
       try {