+++ wallets/leap-extension/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-extension@2.3.0...@cosmos-kit/leap-extension@2.3.1) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/leap-extension
+
+# [2.3.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-extension@2.2.7...@cosmos-kit/leap-extension@2.3.0) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/leap-extension
+
+## [2.2.7](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-extension@2.2.6...@cosmos-kit/leap-extension@2.2.7) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/leap-extension
+
+## [2.2.6](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-extension@2.2.5...@cosmos-kit/leap-extension@2.2.6) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/leap-extension
+
+## [2.2.5](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-extension@2.2.4...@cosmos-kit/leap-extension@2.2.5) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/leap-extension
+
+## [2.2.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-extension@2.2.3...@cosmos-kit/leap-extension@2.2.4) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/leap-extension
+
+## [2.2.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-extension@2.2.2...@cosmos-kit/leap-extension@2.2.3) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/leap-extension
+
 ## [2.2.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-extension@2.2.1...@cosmos-kit/leap-extension@2.2.2) (2023-08-09)
 
 **Note:** Version bump only for package @cosmos-kit/leap-extension