+++ wallets/xdefi-extension/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/xdefi-extension@2.2.0...@cosmos-kit/xdefi-extension@2.2.1) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/xdefi-extension
+
+# [2.2.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/xdefi-extension@2.1.12...@cosmos-kit/xdefi-extension@2.2.0) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/xdefi-extension
+
+## [2.1.12](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/xdefi-extension@2.1.11...@cosmos-kit/xdefi-extension@2.1.12) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/xdefi-extension
+
+## [2.1.11](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/xdefi-extension@2.1.10...@cosmos-kit/xdefi-extension@2.1.11) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/xdefi-extension
+
+## [2.1.10](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/xdefi-extension@2.1.9...@cosmos-kit/xdefi-extension@2.1.10) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/xdefi-extension
+
+## [2.1.9](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/xdefi-extension@2.1.8...@cosmos-kit/xdefi-extension@2.1.9) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/xdefi-extension
+
+## [2.1.8](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/xdefi-extension@2.1.7...@cosmos-kit/xdefi-extension@2.1.8) (2023-08-17)
+
+**Note:** Version bump only for package @cosmos-kit/xdefi-extension
+
 ## [2.1.7](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/xdefi-extension@2.1.6...@cosmos-kit/xdefi-extension@2.1.7) (2023-08-09)
 
 **Note:** Version bump only for package @cosmos-kit/xdefi-extension